Adapt to GOV.UK PaaS
We'd like to be able to maintain the for reliably and painlessly. This
has proven not to be the case, when we're modigying some of the codebase
in order to adapt the project to our platform.

Instead, we're reverting all of these custom changes, which should allow
us to perform `git rebase` without any merge conflicts each time...

We're using the env-map-buildpack to assign Postgres database into our
installation.

We're also forcing some default configuration into place, so the
application behaves on it's own.